Move junitMixin into the Configure interface
Since it is not specific to the Specification DSL only.
