+++ src/main/java/com/greghaskins/spectrum/Configure.java
@@ -7,6 +7,9 @@ import com.greghaskins.spectrum.internal.configuration.BlockTagging;
 import com.greghaskins.spectrum.internal.configuration.ConfiguredBlock;
 import com.greghaskins.spectrum.internal.configuration.ExcludeTags;
 import com.greghaskins.spectrum.internal.configuration.IncludeTags;
+import com.greghaskins.spectrum.internal.junit.Rules;
+
+import java.util.function.Supplier;
 
 public interface Configure {
 
@@ -126,4 +129,15 @@ public interface Configure {
     return new FilterConfigurationChain(new ExcludeTags(tagsToExclude));
   }
 
+  /**
+   * Uses the given class as a mix-in for JUnit rules to be applied. These rules will cascade down
+   * and be applied at the level of specs or atomic specs.
+   *
+   * @param classWithRules Class to create and apply rules to for each spec.
+   * @param <T>            type of the object
+   * @return a supplier of the rules object
+   */
+  static <T> Supplier<T> junitMixin(final Class<T> classWithRules) {
+    return Rules.applyRules(classWithRules, DeclarationState.instance()::addHook);
+  }
 }
